refactor: 패키지 이동 패키지가 이동되면서 public 접근제한자가 추가되었다.
