+++ src/main/java/calculator/type/Divide.java
@@ -1,9 +1,9 @@
-package calculator;
+package calculator.type;
 
 public class Divide extends Operator {
 
     @Override
-    int operate(int num1, int num2) {
+    public int operate(int num1, int num2) {
         if (num2 == 0) {
             throw new IllegalArgumentException("0으로 나눌 수 없습니다.");
         }
