+++ src/main/java/calculator/type/Subtract.java
@@ -1,9 +1,9 @@
-package calculator;
+package calculator.type;
 
 public class Subtract extends Operator {
 
     @Override
-    int operate(int num1, int num2) {
+    public int operate(int num1, int num2) {
         return num1 - num2;
     }
 }
